requirements.txt
dashscope>=0.10.0
firecrawl-py>=1.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
openpyxl>=3.0.9
flask>=2.0.0
requests>=2.26.0
uuid>=1.30
psycopg2-binary>=2.9.1
mysql-connector-python>=8.0.26
# GUI程序不需要额外依赖，tkinter是Python标准库的一部分 